@@ -1,18 +1,23 @@
 {
   "name": "smart",
   "version": "1.0.0",
-  "description": "",
-  "main": "dist/js/index.js",
+  "description": "SMART",
+  "main": "dist/index.js",
   "dependencies": {
-    "lodash": "^4.17.15"
+    "electron": "^8.3.0"
   },
-  "devDependencies": {
-    "electron": "^7.1.1"
-  },
+  "devDependencies": {},
   "scripts": {
     "start": "electron ."
   },
-  "keywords": [],
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/cavineers/smart.git"
+  },
   "author": "",
-  "license": "ISC"
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/cavineers/smart/issues"
+  },
+  "homepage": "https://github.com/cavineers/smart#readme"
 }
