@@ -1,11 +1,7 @@
-let Constants: any;
-
-Constants = {
-    name: {
-        short: "SMART",
-        full: "Super Machine Learning/Artificial Intelligent Robot Tracker"
-    },
-    version: "0.0.1"
-};
-
-export default Constants;
+export default class Constants {
+    public static app_name = class {
+        public static long: string = "Super Machine Learning/Artificial Intelligent Robot Tracker";
+        public static short: string = "SMART";
+    };
+    public static version: string = "0.0.1";
+}
